backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==48.0.0
curl_cffi==0.15.0
distro==1.9.0
fastapi==0.136.1
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
frozendict==2.4.7
google-auth==2.52.0
google-genai==2.0.0
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.12
markdown-it-py==4.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
multitasking==0.0.13
numpy==2.4.4
pandas==3.0.2
peewee==4.0.5
platformdirs==4.9.6
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.2
requests==2.33.1
rich==15.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.7.0
uvicorn==0.46.0
websockets==16.0
Werkzeug==3.1.8
yfinance==1.3.0
